+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/).
 
 - Checkbox no longer becomes disabled when no id is given (fixes [#248](https://github.com/digipolisantwerp/antwerp-ui_react/issues/248))
 - Datepicker label selects datepicker when no id is given
+- Datepicker document event listeners are cleaned up correctly (fixes [#247](https://github.com/digipolisantwerp/antwerp-ui_react/issues/247))
 
 ## v6.0.2 - 2021-06-08